+++ src/spikes/navigator/NavigatorMain.cpp
@@ -75,7 +75,7 @@ void showRoute(bool showLoading = true, navigator::NavigationRefresh refresh = n
     return (navigator::navigationSessionActive() ? tr(STR_NAV_WAITING_GPS) : tr(STR_NAV_GPS_INACTIVE));
   };
   const navigator::NavMapText mapText{tr(STR_NAV_TOTAL_ROUTE), tr(STR_NAV_DURATION), tr(STR_NAV_REMAINING_ROUTE),
-                                      tr(STR_NAV_REMAINING_DURATION)};
+                                      tr(STR_NAV_REMAINING_DURATION), tr(STR_NAV_ARRIVED_STATUS)};
   bool useGray = grayReady;
   bool drawn = navigator::NavScreenRenderer::drawOverview(
       display.getFrameBuffer(), display.getDisplayWidth(), display.getDisplayHeight(), routes.source(), routes.index(),